refactor: split product read concerns into filter and series services
